pspline: fix stdev_cut/max iteration bug
Hi Michael,

Currently line 25 of pspline.py seems to iterate over stdev_cut. This leads to a TypeError when stdev_cut is a float (which generally should be allowed?).

More broadly, I think this might be a bug relative to earlier implementations of this routine, which had this line as for i in range(constants. PSPLINES_MAXITER):, which I think might be correct.

Thank you for creating and maintaining this very useful tool!
-Luke